Includes earlier changes plus first refactor of sidebar
